+++ x/mint/keeper/keeper.go
@@ -75,23 +75,6 @@ func NewKeeper(
 	}
 }
 
-// SetInitialSupplyOffsetDuringMigration sets the supply offset based on the balance of the
-// developer vesting module account. CreateDeveloperVestingModuleAccount must be called
-// prior to calling this method. That is, developer vesting module account must exist when
-// SetInitialSupplyOffsetDuringMigration is called. Also, SetInitialSupplyOffsetDuringMigration
-// should only be called one time during the initial migration to v7. This is done so because
-// we would like to ensure that unvested developer tokens are not returned as part of the supply
-// queries. The method returns an error if current height in ctx is greater than the v7 upgrade height.
-func (k Keeper) SetInitialSupplyOffsetDuringMigration(ctx sdk.Context) error {
-	if !k.accountKeeper.HasAccount(ctx, k.accountKeeper.GetModuleAddress(types.DeveloperVestingModuleAcctName)) {
-		return sdkerrors.Wrapf(types.ErrModuleDoesnotExist, "%s vesting module account doesnot exist", types.DeveloperVestingModuleAcctName)
-	}
-
-	moduleAccBalance := k.bankKeeper.GetBalance(ctx, k.accountKeeper.GetModuleAddress(types.DeveloperVestingModuleAcctName), k.GetParams(ctx).MintDenom)
-	k.bankKeeper.AddSupplyOffset(ctx, moduleAccBalance.Denom, moduleAccBalance.Amount.Neg())
-	return nil
-}
-
 // Logger returns a module-specific logger.
 func (k Keeper) Logger(ctx sdk.Context) log.Logger {
 	return ctx.Logger().With("module", "x/"+types.ModuleName)